refactor: rename routes.ts → route-patterns.ts
The file's main export is ROUTE_PATTERNS (segment constants) plus the
two URL builders derived from them. There's no actual route table here
— that lives in app-router.tsx. The old "routes.ts" name oversold what
was inside and read too close to "app-router.tsx" at a glance;
"route-patterns.ts" accurately describes the contents and visually
separates from the runtime file.
