+++ src/app/loaders/board-set-index-loader.ts
@@ -1,4 +1,4 @@
-import { boardPath } from "@app/routes";
+import { boardPath } from "@app/route-patterns";
 import { getBoardSet } from "@features/board";
 import { m } from "@paraglide/messages.js";
 import { data, redirect, type LoaderFunctionArgs } from "react-router";
